visualisasi/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
python-dotenv
plotly
sqlparse
